style: fix formatting using prettier

+++ interfaces/solidly/IBaseV1Minter.sol
@@ -3,7 +3,6 @@
 pragma solidity ^0.6.0;
 pragma experimental ABIEncoderV2;
 
-
 interface IBaseV1Minter {
-      function update_period() external returns (uint);
+    function update_period() external returns (uint256);
 }
