+++ interfaces/solidly/IveDist.sol
@@ -5,7 +5,9 @@ pragma experimental ABIEncoderV2;
 
 // NOT SURE WHAT THSI CONTRACT DOES, but you claim stuff
 interface IveDist {
-    function claimable(uint _tokenId) external view returns (uint);
-    function claim(uint _tokenId) external returns (uint);
-    function claim_many(uint[] memory _tokenIds) external returns (bool);
+    function claimable(uint256 _tokenId) external view returns (uint256);
+
+    function claim(uint256 _tokenId) external returns (uint256);
+
+    function claim_many(uint256[] memory _tokenIds) external returns (bool);
 }
